requirements.txt
annotated-types==0.7.0
anyio==4.9.0
anthropic==0.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
dpath==2.1.6
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipython>=5.3.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpickle>=1.4.1
jsonpointer==3.0.0
langchain-core==0.3.58
langchain-openai==0.3.16
langgraph==0.4.1
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.42
matplotlib==3.8.4
networkx==3.3
openai>=1.68.2,<2.0.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.2
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
pyvis==0.3.2
pytest==8.3.5
pytest-timeout
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
WeasyPrint==62.3
xxhash==3.5.0
zstandard==0.23.0

# Add missing langchain dependencies
langchain-google-genai>=0.0.1
langchain-core>=0.1.0 # Specify a reasonable minimum version
pydantic>=2.0.0 # Specify a reasonable minimum version

# New langchain-cerebras dependency
langchain-cerebras>=0.0.1 # For Cerebras model integration

# New sentence-transformers dependency
sentence-transformers

# New chromadb dependency
chromadb

# New fastcdc dependency
fastcdc

# New pytest-asyncio dependency
pytest-asyncio

# For API server
fastapi
uvicorn[standard] # [standard] includes common extras like websockets, etc.

# For interactive UI (Level 5 Transparency)
streamlit
